cs-300: Extract the ML model to a separate binary
Eventually, we want the MCUBoot (and the rest of TF-M) to handle the ML
model in the same way as other components. To achieve that, the ML model
component will be kept in flash during boot, for the MCUBoot to validate
the image, and perform the swap operation if needed.

Since the Ethos NPU is not allowed to access flash, the model will be
copied back to DDR at runtime. This is why the model is still kept in
the DDR memory region in the linker script.

Signed-off-by: Filip Jagodzinski <filip.jagodzinski@arm.com>