ml-update-demo: Make the model update more visible
Update from a broken model to a good one.

Add a modified tflite file that produces no results at runtime. This
file can be used to build an application that is fully functional, but
the ML inference is unsuccessful in detecting any keyword. Applying a
OTA ML model update in this state is very visible since the updated
model does work as expected (detects keywords from audio sampes).

Signed-off-by: Filip Jagodzinski <filip.jagodzinski@arm.com>